Add options for fine-tuning nice/ionice local/remote usage
Sometimes the nice/ionice programs are not available on one
of the sides. In that case, the four new options allow for
fine-grained selection of the commands to run or not
(and the niceness levels).

The four options are:
* --local-nice  (default to the --nice value if unset)
* --remote-nice (default to the --nice value if unset)
* --local-ionice  (default to the --ionice value if unset)
* --remote-ionice (default to the --ionice value if unset)

The options are also use for the file deletions handled under
the --cleanup option.