Ensure the list of link-dest directories contains backups for the host
The issue with the --others option is that when a host hasn't been
backed up for a while, the list of directories for the --link-dest
option will only contain directories for "others". This happens
because all the link-dest candidates are sorted by date.

This patch falls back to putting backups for the host itself
first when the link-dest list contains no older backups.