Add a script to fatpack dumbbackup
My first attempt at fatpacking produce a script that wouldn't run
on a bare Perl installation.

This packer recursively scans the unpacked script and the lib/
directory for dependencies, and has a list of optional modules
to skip.

It also removes all the documentation for the dependencies,
to save space.