introduce a new framework for building CLI tools
This is an idea I've had for over a decade now: build a command-line
tool by applying roles to the code for a single command.

For example, a role can add support for specific options, and with
clever use of lazy attributes and method modifiers, easily provide a
shared set of options to subcommands.

This implementation will eventually move out of the dumbbackup
repository, but I'd like to experiment with it first.

I'm calling this approach "Role Your Own" CLI interface, or RYO.