save the backup in progress in a separate directory
When the backup fails (e.g. because of a network timeout),
the unfinished backup is not left among the complete ones
(which could cause data loss if that incomplete backup is
kept and a complete one removed).

The issue is that the target directory must not exist,
otherwise instead of:

    .inprogress -> 2025-11-21

we'll get:

    .inprogress -> 2025-11-21/.inprogress